+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for tasty-flaky
 
+## Unreleased
+
+* Add `flakyTestWithRetryAction`. Retry logic can now be based on test results.
+
 ## 0.1.1.0 -- 2024-12-19
 
 * Explicit support and testing for GHC 9.12